added directives for updating or deleting user account

+++ public/app/controllers/register.js
@@ -2,21 +2,23 @@
   'use strict';
 
   angular.module('like.register', [])
-  .controller('registerCtrl', ['$scope', '$http', '$location', function ($scope, $http, $location) {
-    $scope.register = function (useObj) {
-      $location.path('/dashboard');
-      return $http({
-        method: 'post',
-        url: '/api/user/create',
-        data: useObj
-        //useObj: {username: username, password: password}
-      }).then(function (data) {
-        sessionStorage.setItem('userId', data.data.userId);
-        return data.data.userId;
-      }).catch(function (err) {
-        return err.status;
-      });
-    }; //close register
+  .controller('registerCtrl', ['$scope', '$http', '$location', 'authService', function ($scope, $http, $location, authService) {
+    $scope.userObj = {};
+    $scope.emailCheck = function (str) {
+      return /^\s*[\w-_\.\+]+@[\w-]+\.[\w-\.]+$/.test(str);
+    };
+    $scope.register = function (userObj) {
+      if (userObj.password === userObj.confirm) {
+        return authService.register(userObj)
+        .then(function (data) {
+          sessionStorage.setItem('userId', data.data.userId);
+          return data.data.userId;
+        }).catch(function (err) {
+          return err.status;
+        });
+      }
+    };
+    // }; //close register
 
     }]); //close controller def
 })();
